various fixes -refactored meroshare import code -meroshare to portfolio(purhases), sales (sales), portfolio_summary(cummulative) -enhancements portfolio edit and new forms -fixed migrations -calculations of broker_commission, sebon_commission for purchases via Broker'

+++ database/migrations/2020_12_04_095438_create_news_comments_table.php
@@ -19,7 +19,7 @@ class CreateNewsCommentsTable extends Migration
             $table->foreignId('user_id')->constrained('users');
             $table->string('title');
             $table->string('comments');
-            $table->smallint('rating');
+            $table->integer('rating');
             $table->timestamps();
         });
     }
